rename file to mirror main class name in file
